pdf-similarity/requirements.txt

nltk==3.6.5

python-Levenshtein==0.12.2

textdistance==4.2.1

PyPDF2==1.26.0

scikit-learn==0.24.2

PyMuPDF==1.18.15

scikit-image==0.18.3

opencv-python==4.5.3.56

numpy==1.20.3

pandas==1.3.3

streamlit==0.86.0

Pillow==8.3.2

image_text_similarity==1.0.1
